Shoko - Compile-time UI Framework [Rendering Tech Demo]
Небольшая оговорка: Реализованы не все RHI, конкретно сейчас работает на RayLib

+++ Source/Shoko/Core/Math.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <corecrt_math.h>
+#include <cmath>
 
 namespace Shoko
 {
@@ -11,7 +11,7 @@ namespace Shoko
         template <typename T> static T Min(T A, T B) { return A < B ? A : B; }
         template <typename T> static T Max(T A, T B) { return A > B ? A : B; }
 
-        static float Sqrt(float A) { return sqrtf(A); }
+        static float Sqrt(float A) { return sqrt(A); }
 
         static constexpr float PI = 3.14159265358979323846f;
     };
